% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BioDivExpt.1}
\alias{BioDivExpt.1}
\title{Annual-plant diversity experiment 1}
\format{
\code{BioDivExpt.1} is a list with the following four elements:
\describe{
  \item{InitDens}{Initial seed densities (seeds sown per pot) of each species in each pot, organized in a data frame with 45 rows (pots) and 5 columns (species). The remaining three elements (FinalDens, FinalYield, and GermDens) follow the same format, with the pots and species arranged in the same orders as in InitDens.}
  \item{FinalDens}{Final plant densities (plants survived per pot) of each species in each pot.}
  \item{FinalYield}{Final plant biomass (grams per pot) of each species in each pot.}
  \item{GermDens}{Seedling densities (seedlings emerged per pot) of each species in each pot.}
}
}
\usage{
BioDivExpt.1
}
\description{
A dataset of a diversity experiment using annual plants. Five species (\emph{Arthraxon hispidus}, \emph{Bidens pilosa}, \emph{Digitaria ciliaris}, \emph{Kummerowia stipulacea}, and \emph{Persicaria longiseta}) were sown solely or in mixtures of two, three, or four species in planting pots. In this experiment (\code{BioDivExpt.1}), a total of 24 seeds were sown per pot, while in another experiment (\code{BioDivExpt.2}), a total of 72 seeds were sown per pot. The number of emerged seedlings were counted after 2 weeks of sowing. After a growing season, the number of live plants were counted. Plants were then  harvested, dried, and weighed.
}
\author{
Shinichi Tatsumi
}
\keyword{datasets}
