% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densize.R
\name{densize}
\alias{densize}
\title{Partitioning biodiversity effects into density and size components}
\usage{
densize(init.dens, final.dens, final.yield, germ.dens)
}
\arguments{
\item{init.dens}{A matrix or data frame consisting of initial plant density (the number of plants or seeds sown per area).
Rows are plots and columns are species.
Elements should be filled with 0 for plots without focal species.
The arguments \code{final.dens}, \code{final.yield}, and \code{germ.dens} follow the same format, with the plots and species arranged in the same order as in \code{init.dens}.}

\item{final.dens}{A matrix or data frame consisting of final plant density (the number of plants survived).}

\item{final.yield}{A matrix or data frame consisting of final yield (yield per area).
Yield can be in any unit; e.g., biomass (g), volume (m3).}

\item{germ.dens}{Optional. A matrix or data frame consisting of germination density (the number of seedlings emerged per area).
When provided, germination components are calculated in addition to density and size components.}
}
\value{
The \code{densize} function returns a matrix with rows as plots and columns as types of components. Rows representing monoculture plots, for which biodiversity cannot be defined, are removed.
\itemize{
 \item{When the argument \code{germ.dens} was NOT provided, the function returns (1) density-mediated complementarity effects (\code{dens.compl}), (2) density-mediated selection effects (\code{dens.selec}), (3) size-mediated complementarity effects (\code{size.compl})}, and (4) size-mediated selection effects (\code{size.selec}).
 \item{When the argument \code{germ.dens} was provided, the function returns (1) germination-mediated complementarity effects (\code{germ.compl}), (2) germination-mediated selection effects (\code{germ.selec}), (3) post-germination density-mediated complementarity effects (\code{pg.dens.compl}), (4) post-germination density-mediated selection effects (\code{pg.dens.selec}), (5) size-mediated complementarity effects (\code{size.compl})}, and (6) size-mediated selection effects (\code{size.selec}).
}
}
\description{
Function \code{densize} (DENSIty and SIZE) partitions the net, complementarity, and selection effects of biodiversity into additive components that reflect diversity-induced changes in plant density and size.
}
\examples{
dat  <- BioDivExpt.1
res1 <- densize(dat$InitDens, dat$FinalDens, dat$FinalYield)
res2 <- densize(dat$InitDens, dat$FinalDens, dat$FinalYield, dat$GermDens)
# res1: Density and size components
# res2: Germination, post-germination density, and size components
#  Note that the input dataset (BioDivExpt_1) has 45 rows, while
# each output object has 30 rows. This is because the densize
# function has removed rows representing monocultures (1 to 15),
# for which biodiversity effects cannot be defined.
}
\author{
Shinichi Tatsumi
}
