% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densize.R
\name{densize}
\alias{densize}
\title{Partitioning biodiversity effects into density and size components}
\usage{
densize(init_dens, final_dens, final_yield, germ_dens)
}
\arguments{
\item{init_dens}{Matrix or data frame consisting of initial plant (seed) density. Rows are plots and columns are species. Elements should be filled with 0 for plots without focal species. The same applies to \code{final_dens}, \code{final_yield}, and \code{germ_dens}.}

\item{final_dens}{Matrix or data frame consisting of final plant density. Note that the plots and species must be in the exact same order as \code{init_dens}. The same applies to \code{final_yield} and \code{germ_dens}.}

\item{final_yield}{Matrix or data frame consisting of final yield.}

\item{germ_dens}{Optional. Matrix or data frame consisting of germination density (e.g., seedling counts). When provided, germination components are calculated in addition to density and size components.}
}
\description{
Function \code{densize} (DENSIty and SIZE) partitions the net, complementarity, and selection effects of biodiversity into additive components that reflect diversity-induced changes in plant density and size.
}
\examples{
aaa
}
\author{
Shinichi Tatsumi
}
